pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "psycopg2-binary>=2.9.10",
]
